+++ tests/unit/app/templates/fakeaggregate_index.php
@@ -1,0 +1,11 @@
+<html>
+<head>
+</head>
+<body>
+<!-- Plates template -->
+<?php echo $view['results']; ?>
+<?php echo $view['values'][1]['testkey']; ?>
+<?php echo $view['values'][0]['testreadmodel']; ?>
+<?php echo $view['preindex']; ?>
+</body>
+</html>