+++ docs/whats_new.rst
@@ -1,5 +1,18 @@
 .. _WhatsNewAnchor:
 
+What's new in version 3.1.0 (2019-04-30)
+========================================
+
+* Adds aggregate root controllers for easier usage of aggregate-based functionality.
+
+* Adds asynchronous functionality to the Event Dispatcher.
+
+* Adds asynchronous commands in non-Swoole environments using ReactPHP/Symfony Process.
+
+* Adds a Command Runner to make running commands independent of the PHP environment (Swoole or non-Swoole).
+
+* Updates the documentation.
+
 What's new in version 3.0.0 (2019-04-16)
 ========================================